Library

LiteQTL.cpurun

LiteQTL.cpurun

LiteQTL.get_geno_data

LiteQTL.get_pheno_data

LiteQTL.gpurun

LiteQTL.scan

LiteQTL.scan

LiteQTL.cpurun — Method

cpurun(Y, G, n, export_matrix)

Arguments:

Y : a matrix of phenotypes

G : a matrix of genotypes

n : the number of individuals

export_matrix : a boolean value that determines whether the result should be the maximum value of LOD score of each phenotype and its corresponding index, or the whole LOD score matrix.

Output:

returns the maximum LOD (Log of odds) score if export_matrix is false, or LOD score matrix otherwise.

LiteQTL.cpurun — Method

cpurun(Y, G, X, n, export_matrix)

Arguments:

Y : a matrix of phenotypes

G : a matrix of genotypes

X : a matrix of covariates

n : the number of individuals

export_matrix : a boolean value that determines whether the result should be the maximum value of LOD score of each phenotype and its corresponding index, or the whole LOD score matrix.

Output:

returns the maximum LOD (Log of odds) score if export_matrix is false, or LOD score matrix otherwise.

LiteQTL.get_geno_data — Method

get_geno_data(file, datatype)

returns the genotype data. Will skip every other column because genotype probability is duplicated.

LiteQTL.get_pheno_data — Method

get_pheno_data(file, datatype; transposed)

returns the phenotype data. If transposed=true, then the data will be transposed.

LiteQTL.gpurun — Method

gpurun(Y, G, n)

Arguments:

Y : a matrix of phenotypes

G : a matrix of genotypes

n : the number of individuals

Output:

returns the maximum LOD (Log of odds) score

LiteQTL.scan — Method

scan(Y, G, n; export_matrix, usegpu)

This function will run genome scan without covariates.

Arguments:

Y : a matrix of phenotypes

G : a matrix of genotypes

n : the number of individuals

export_matrix : a boolean value that determines whether the result should be the maximum value of LOD score of each phenotype and its corresponding index, or the whole LOD score matrix.

usegpu : a boolean value that indicates whether to run scan function on GPU or CPU. Default is false, which runs scan on CPU.

Output:

calls cpurun function if usegpu=false, otherwise, calls gpurun

LiteQTL.scan — Method

scan(Y, G, X, n; export_matrix, usegpu)

This scan function will run

Arguments:

Y : a matrix of phenotypes

G : a matrix of genotypes

X : a matrix of covariates

n : the number of individuals

export_matrix : a boolean value that determines whether the result should be the maximum value of LOD score of each phenotype and its corresponding index, or the whole LOD score matrix.

usegpu : a boolean value that indicates whether to run scan function on GPU or CPU. Default is false, which runs scan on CPU.

Output:

returns the maximum LOD (Log of odds) score if export_matrix is false, or LOD score matrix otherwise.