Library

LiteQTL.cpurun

LiteQTL.get_geno_data

LiteQTL.get_pheno_data

LiteQTL.scan

LiteQTL.get_geno_data — Method

get_geno_data(file, datatype)

returns the genotype data. Will skip every other column because genotype probability is duplicated.

LiteQTL.get_pheno_data — Method

get_pheno_data(file, datatype; transposed)

returns the phenotype data. If transposed=true, then the data will be transposed.

LiteQTL.scan — Function

scan(Y, G)
scan(Y, G, X; maf_threshold, export_matrix, usegpu, lod_or_pval, timing_file)

This scan function will run

Arguments:

Y : a matrix of phenotypes

G : a matrix of genotypes

X : a matrix of covariates

n : the number of individuals

export_matrix : a boolean value that determines whether the result should be the maximum value of LOD score of each phenotype and its corresponding index, or the whole LOD score matrix.

usegpu : a boolean value that indicates whether to run scan function on GPU or CPU. Default is false, which runs scan on CPU.

Output:

returns the maximum LOD (Log of odds) score if export_matrix is false, or LOD score matrix otherwise.

LiteQTL.cpurun — Function

cpurun(pheno, geno)
cpurun(pheno, geno, X; maf_threshold, export_matrix, lod_or_pval, timing_file)

Arguments:

Y : a matrix of phenotypes

G : a matrix of genotypes

n : the number of individuals

export_matrix : a boolean value that determines whether the result should be the maximum value of LOD score of each phenotype and its corresponding index, or the whole LOD score matrix.

Output:

returns the maximum LOD (Log of odds) score if export_matrix is false, or LOD score matrix otherwise.